refacto(air/layout): move errors to their own file
